pyproject.toml
[tool.poetry]
name = "viecpro-backend"
version = "0.1.0"
description = "Backend for the VieCPro project."
authors = ["Matthias Schlögl <matthias.schloegl@oeaw.ac.at>", "Gregor Pirgie <gregor.pirgie@oeaw.ac.at>", "Peter Andorfer <peter.andorfer@oeaw.ac.at>", "Stefan Resch <stefan.resch@oeaw.ac.at>", "Richard Hadden <richard.hadden@oeaw.ac.at>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
whitenoise = "^5.2.0"
sentry-sdk = "*"
dj-database-url = "^0.5.0"
gunicorn = "^20.0.4"
acdh-django-charts = "^0.5.3"
django-allow-cidr = "^0.3.1"
mysqlclient = "^2.0.3"
django-extensions = "^3.1.3"
django-csp = "^3.7"
apis-bibsonomy = "0.4.1"
apis-ampel = "^0.2.0"
pylint = "^2.6.0"
jupyter = "^1.0.0"
jupyter-tabnine = "^1.2.3"
black = "^22.12.0"
flake8 = "^6.0.0"
celery = "^5.3.4"
python-Levenshtein = "^0.23.0"
django-celery-results = "^2.5.1"
django-debug-toolbar = "^4.2.0"
Django = "3.2.18"
psycopg2 = "^2.9.7"
traitlets = "<=5.9.0"
jupyter-server = "^2.7.3"
typesense = "^0.17.0"
redis = "^5.0.1"
pandas = "^2.0"
djangorestframework = "^3.14.0"
django-filter = "^2.4.0"
django-autocomplete-light = "~3.8.2"
django-cors-headers = "^3.5.0"
django-crum = "^0.7.9"
django-crispy-forms = "^1.10.0"
django-gm2m = "^1.1.1"
django-guardian = "^2.3.0"
django-leaflet = "^0.27.1"
django-reversion = "^3.0.8"
django-tables2 = "^2.3.3"
djangorestframework-csv = "^2.1.0"
djangorestframework-xml = "^2.0.0"
jmespath = "^0.10.0"
jsonschema = "^3.2.0"
lxml = "^4.6.2"
convertdate = "^2.3.0"
rdflib = "~6.1.1"
drf-spectacular = "~0.26.1"
tablib = "^3.0.0"
regex = "^2020.11.13"
requests = "^2.25.0"
SPARQLWrapper = "^1.8.5"
unicodecsv = "^0.14.1"
django-model-utils = "^4.1.1"
django-summernote = "^0.8.11"
djangorestframework-jsonschema = "^0.1.1"
django-admin-csvexport = "^1.9"
tqdm = "*"


[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.core.masonry.api"